init Procfile, Procfile.dev:
Created Procfile, Procfile.dev to save the settings and seperated it
into two files, one for production deployment and the other for
development deployment. For the web deployment the guincorn will be
used, and celery would be used for distributed task queue features to give
users more efficient user experience.

Also added honcho and gunicorn packages for deploying server. Honcho is
to manage Procfile based applications to manage multiple processes
efficiently. Gunicorn is a Python WSGI HTTP server for UNIX.