init serializers:
Initiated api serializers for User and Post models. Each serializer is
to create APIview for various purposes. UserModelSerializer is based on
User model and PostModelSerializer is based on Post model.

In PostModelSerializer, username field is added due to the fact that the
Post model is dependent on User model. Because object of User model
creates Post model's object as One to Many relationship.

Also each model's so called 'id' is different. Based on the Model, User
model's id is "id" and Post model's id is "post_id". Got to be careful
about this.