created apis application:
Since this project's purpose is to create an ios app, API is one of the
core function. In consideration of the importance of API and it's
number, it is worth to be seperated as a single independent application.

In this application serializers, APIView and API related urls are going
to be included. Any furthur api related files must be included in this
folder.