install djangorestfreamwork-jwt:
This package is JWT method for authentication provided for django rest
freamework. JWT method is an appropriate authentication way to interac
with iOS apps. In web browser, usually authentication tokens are stored
as sessions and cookies. But iOS app can't store it with session method.
So it needs an JSON token to authenticate and verify the user. This
package easily enables it.

http://getblimp.github.io/django-rest-framework-jwt/