init rest_framework_jwt urls:
According to the djangorestframework-jwt document, it says to connect
specific urls to each token functions.

The /api/token/obtain/ to obtain_jwt_token is for authentication when
the request is made as POST method and returns the user's unique token.
This it deemed as login process. Since the token is returned, iOS client
should store this token for further authentication.

The /api/token/verity/ to verity_jwt_token is for validating the user's
authority. When the token is sent and the server checks the validity of
the user and gives back only the authorized resources.

The /api/token/refresh/ to refresh_jwt_token is for refreshign the
token value. Because server expires the toeken of the user,so user may not be able to authenticate and verified with formerly given toekn. This function is to refresh and create a branch new token for a user.