update post api views:
updated post.py for post api views. Added PostUpdateAPIView and
PostDeleteAPIView. Each inherits UpdateAPIView and DestroyAPIView which
are the genericviews provided by django rest framenwork. UpdateAPIView
is to update the instance through the api access and DestroyAPIView is
to delete the instance. Right now each view takes the
'PostDetailSerializer' serializer class. This can be changed in near future.