update post api views: perform_***
Added perform_create and perform_update functions in create and update
view. This function perform designated function when saving or updating a object instance.

In this case, it takes an user from the request and save it in
serializer object instance. So it automatically put user and username
when create or update the post.