update post apis:
There was a problem with urls of finding pks of slugs of the post
instance. I found out the regex in urls were set as word strings not as
number(primary key). So changed into -d in regex.

Also updated some minor changes. Unnecessary fields in serializers were
removed. When creating a post instance, the user information is
automatically passed in to the instance. So no longer need an username
or user field to create a post.

Also deleted former post_id field and added pk field to explicitly show
and differentiate between different posts. Voila!