init permissions.py and adjust it to views:
The actions in views of posts needs to be authenticated.

For example,
lists of post can be viewed by anyone. But to create a post there has to
be a user to create. In this case I'm not letting anonymous user to
create a post.

Once the post is created, it can be edited only by the owner of the post
which is the user who created the post. So I created IsOwnerOrReadonly
for the updateapiview. So in this case, any users can read the post in
updating url, but cannot edit it. Only post owner can do it.