init user create view and serializer:
Initicated UserCreateSerializer and UserCreateAPIVIew. In serializer,
the password can't be read but writable. When creating the user,
password should be written. But when the user data is returned, the raw
password shouldn't be shown. And in this state, password is shown as
hashed value.

In UserCreateAPIView, it refers to UserCreateSerializer serializer and
inherits CreateAPIView.