update UserCreateSerializer, create method added:
Before this, the UserCreateAPIView doesn't actually passed the password
for that user. It was saving in the user model but didn't set the
password correctly.

In create method, username, password and phonenumber is get from the
validated_data and passed into respective value. Then it creates User by
passing in username in username and phonenumber in phonenumber.

Once the user instance in created, it should set the password with
.set_password method by passing in password. Then saves the instance.