init UserLoginAPIView: without validation
Created UserLoginSerializer and UserLoginAPIView.

In UserLoginSerializer, the token and username field is added with
CharField as a custom field to be used for serialized data. The token
here is for authentication. Also password should be set as write only.

In UserLoginAPIView, the permission classes would be set as AllowAny
because anyone should be able to login. The serializer here is linked to
UserLoginSerializer.

If the requested data is POST method and when it is passed into
serializer and it is valid, the view returns Response of HTTP status
200. Else, when the data is not valid and raises error, it returns
Response of HTTP status 400 bad request.

Until now, a basic process of login is written but it doesn't check the
validity of the account. This should be added soon.