update UserLoginSerializer: added validation
Once the user tries to login, the data has to be checked if it is
valid one or not.

Once the request is made, the validate method runs. It gets username and
password from the data. If there is no username input it returns error
message that the user should input username. Else, it finds the matching
username in existing user objects. If there is a matching username, it
saves the user into user_obj variable. If there is no matching username
in user objects, it raises error. Then, checks the password of user_obj
with the input password are matching. If not, raises an error.

Finally if username and password matches the existing user object, the
view adds token value in the data["token"] and return in for furthur
authentication that client can user. For now, the token is not real one.