update production.py and base.py for s3 settings:
settings of static files path in base.py was set to local dist's static
folders. So I changed this to aws s3's root. Also in production.py
setting, files storate s3 name was set as fastube because the setting
code was copied and pasted in here. Changed to sooljottagrae.