update serializers of user and post:
As the User and Post model changed, the fields in meta had to be
changed. username is no longer neede so changed to email.

Also validation of email and password was included. Validating email
whether it is already used one or not. Also taking password confirm for creating an user through the API.