update social auth: facebook included
In base settings, FasebookOAuth2 is included as authentication backends by
importing from social auth package.

To access the social login, we have to get into the specific url which
is including backend after the login url. But due to the missing '$'
word in regexes in /login/, /signup/, /logout/, django didn't catch the
social auth url. So I added '$' after each url.

But I wanted to add the social auth url in users.urls to make the social
auth url to be as /user/login/facebook. But it turns out
an error of namespace error that social is not defined one. Therefore I
included social url info in sooljottagrae's base url file. So the
facebook login url is /login/facebook.

I hope to figure out this mismatching url path.