update settings for authentication:
Changed url of facebook login to /user/login/facebook by declaring
SOCIAL_AUTH_URL_NAMESPACE = 'user:social' because the users.url's name
space is 'users' so added the name space'usere' in front of 'social'.

Also revised LOGIN_URL path to /user/login/ which is the login url.
Based on this setting, when using the @login_required decorators, it
redirects to the this setting's path.