Init post hyperlink urls:
HyperlinkedIdentityField, which shows an url of the specific
objects was added in serializer.py. This looks into post detail view and
get the url of each post. Through this process it looks up a 'pk' to get
the specific number of the post. Then stores the url as post_detail_url.

This url is now shown in post lists, and detail api.