update post serializers:
In post list serializer, created an method get_user to get the user
information and passed the returned result to user field as a
SerializerMethodField then put it in fields context.

In post detail serializer, created a get_image and get_user method then
put it in filed as a serializerMethodField. Now post detail api shows
the image url and and user information.

Also created image field in POstcreateupdateserializer to let users able
to upload image through apis. But still the image file is not uploaded
to proper path.