init comment apis:
Created an api for comments to enaable to CRUD comments. Right now, I
created Comment Serializer which is an base serializer. Then created
comment detail and list api views and connected with comment serializer.

Also added comments url in urls file. But right now in the comment list,
it doesn't show any relationship with posts. The list here is just
showing all the list of comments that are existings but not showing any
information of post.

So the comment information is going to be added in the post detail api
later on. This just single comment list is not very useful.