update comment apis: comment create api
Created a comment create serializer which only takes content for fields.
Then in comment create api view, when it takes the post request, it gets
post objects through the url's pk which refers to the specific post that
the user intended to create comment on it.

Then in its post object, comment is created by taking the content that the user wrote and the user information from the request.

When comment is successfully created, the view returns Response of status code with the user's nickname and the comment's content.