Update comment edit and detail apis:
Comment edit serializer was created for commentEditAPiVIew. It is
composed of RetrieveAPIView and update & destroy model mixins. These two
mixins allows the view to update the object with put method and destroy
it with delete method. The url was set after the number of comment id by
adding /edit.

For CommentDetailAPIView, it uses CommentSerializer and looks up for the
comment id as 'id' in the url. The view overrides get_queryset method
because the queryset itself is determined by the instance of post
because the comment is depedent on the post model. So the method finds
the specific post instance with 'pk' and put the instance in the
variable post. Then it finds the comment querset of the post instance.

The urls were all included in the post url, since most of the comment
views needed an information of the post instance. Therefore, it is much
more natural that the comments urls to be added after the post
url(information).

But the comments list that only shows things subordinate to the post is
not created yet.