update comment lists: seperated list views
Former commentlist view was seperated in two different kind of list
views. The first one is general list which shows all the comment
instances and have a function to search the comments with filterint
method. The other on is specific comment list which shows comments that
is dependent on the specific posts. So user can access to comment lists
naturally by putting comments/ after the specific post url. This list
doesn't have a search function.