update comment api:
To add the detail url of the comment instance in the comment specific
lists, general lists and detail, I added comment_detail_url field of
HyperlinkedIdentityField which refers to apis:comments:detail view and
looksup "id" to identify the comment instance.

Also the comment detail serializer was created to connect it with
comment detail api view.

So the comment's retrieve(general), update, delete can be performed with
apis:commments urls.
On the other hand, comment's create, retrieve(specific) can be peroemed
with apis:posts:comments. Which actions are dependent on post instance.