update post model: functions related to imagefield
What I'm doing here is to rename the image file's name that users upload
when creating a post. When user write the post, the image's filename
must be changed to those of unique one. If it isn't there probably would
be a conflict between image files. To prevent this to happen, it is good
idea to rename the file's name to the time when the post was created. So
the file name format would be 'year''month''day' + 'microsecond. So the
set_filename_format function performs this process.

Then the user_directory_path function performs creating specific folder
to store the image file in terms of date. It depends on the year, month,
day and user's info such as email and then the unique filename.

Finally in the post model, the imagefield itself gets upload_to
parameter to set the directory path when the image was uploaded. So in
this case it takes user_directory_path function to have the file change
the filename and store in the desired path.

All this idea of changing filename and making specific path is from
LeoHeo, best student of fastcampus's web programming school. This is his
blog address that I referred to:
http://hjh5488.tistory.com/12