Init tags api:
Out model have three different types of tags: AlcoholTag, FoodTag and
PlaceTag. These three tags are all categorized as tags but to be able to
manage them well, three different tag kinds are seperated in three
models.

According to the initial design of tags, the apis are seperated. So the
each model's serializers were created and then each object's list views
were created.

Soon, the post informations would be added. Also the CRUD of tags have
to be made. But since the service have fixed and preset tags are going
to be suggested, the creation and editing tags will be limited to app's
superuser.