update three tag serializer: include post info
In each tag serializer, the post objects that have relationship with
each tag instance have to be shown. Since it is an manytomany
relationship with post.

But in the tag.py, I tried to import one of post serializer but couldn't
import it. I looked it up on the google, and found out there would be an
circular import error because of the relationship between post and tags.

So I just created simple TagPostSerializer that serializes post
imformation but only includes post id. I tried to include url of each
post instance with HttpIdentityField, but an error occured regarding
context missing. So this time I just took it out temporarily, but hope
to add it soon.

Anyway now in post detail and tag detail, user can see each instance's
following(relating) instances.