edit tag serializers, comment model:
There were difficulty in brining each instance's absolute api urls in
the related model instance's detail view.

For example in one post instance, it shows all the following comment instances. However it is supposed to show the comment's url in the instance serialized data, due to the hyperlinkedIdentity field's error(related to request context), it can't bring it unless the views were modified.

Therefore as a stopgap to solve this problem, I added
get_absolute_api_url method within the comment model. To be able to
bring the api url with the model method. I included it only in the serializer
that is called in other related serializer. In the independent detail
serializer, it just uses hyperlinkedidentity field.

The tags instance's urls would be solved same way as this one.