update post models, tag model and serializers:
In the post model, getting absolute api url method was created. This
method is now used in the related serailization within the tags
serializers.

Also in the three tag models, api url method was created too, then used
in related serializer(post).

The url is currently hardcoded as https://sooljotta.com but this have to
be refactored in future release.