update app applications: included JWT settings
Created jwt.py settings file in the setting's partial directory. These
are all the jwt setting contents so from now on, all the jwt
authentication related settings are handled in this file.

Also the django application secret was took out now... So late but it is
no longer public. Have to change this sercret key since it is already
uploaded to github.