update post update and delete view:
The post's update and delete was seperated before but integrated both
function in the url name edit and changed the two views into one which
called PostEditAPIView. Therefore in this edit view, user can update or
delete the post with the request method. With a put method, user can
update it and with delete method the user can destroy the instance.

Also comment edit view was not included in the module. So I added it in
it.