update post, tag serializers:
Seperated post update serializer from the createupdate serializer. When
creating post, user only have to input the id of the desired tags. But
when updating the post, the serializer doesn't returns meaningful
information of the tags.

So in the new update serializer, the tag instance's information is
serialized and explicitly shown in the api as json with the parameter
and the value of it.