init mytagslistapiview:
This api view is to show all the list of posts which contains anytags
that the user follows.

So in the view, it takes the user instance through the pk which is
included in the url. It finds the user instance then takes tags that
user is following. In this case, it gets three different kinds of tag
models.

After that, the view takes all the related post instances from the
extracted tags. Each post instances are saved in the new list
'post_queryset'.

Now post_queryset contains all the posts that is related to tags which
user followed.

In the url.py, my tags view is placed after the user detail url. To make
it easy to know the post lists that is related to specific user's
following tags.