update user model, serializer: avatar field
In the user model, upload directory can't get the user instance because
when creating an user instance, the user instance doesn't exist. So
didn't include username in the file formating function.

Also in the seraiziser, avatar had to be passed into the User parameter.
So it can contain the avatar information.